retina-oct-demo-app/pyproject.toml
[project]
name = "retina-oct-demo-app"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "streamlit>=1.28.0",
  "torch>=2.0.0",
  "torchvision>=0.15.0",
  "numpy>=1.24.0",
  "Pillow>=9.5.0",
  "matplotlib>=3.7.0",
  "plotly>=5.15.0",
  "opencv-python-headless>=4.8.0",
  "scikit-learn>=1.3.0",
  "segmentation-models-pytorch>=0.3.3",
  "watchdog>=6.0.0",
]

[tool.setuptools]
packages = []
